requirements.txt
click>=8.1.7

# Core AI Framework
crewai>=0.130.0
crewai-tools>=0.47.1

# API Framework
fastapi>=0.110.3
uvicorn>=0.29.0

# Google AI SDK
google-ai-generativelanguage>=0.6.4
google-api-core>=2.10.0
google-api-python-client>=2.131.0
google-auth>=2.29.0
google-auth-httplib2>=0.2.0
google-cloud-aiplatform>=1.53.0
google-cloud-bigquery>=3.23.1
google-cloud-core>=2.4.1
google-cloud-resource-manager>=1.12.3
google-cloud-storage>=2.16.0
google-crc32c>=1.5.0
google-generativeai>=0.5.4
google-resumable-media>=2.7.0
googleapis-common-protos>=1.63.0

# Core Libraries
Jinja2>=3.1.4
jsonschema>=4.22.0
numpy>=1.26.4
pandas>=2.2.2
pillow>=10.3.0

# Langchain Ecosystem
langchain-core>=0.1.52
langchain-community>=0.0.38
langchain-openai>=0.1.7
langsmith>=0.1.67

# AI Models
openai>=1.30.5
onnxruntime>=1.22.0

# Telemetry
opentelemetry-api>=1.30.0
opentelemetry-exporter-otlp-proto-common>=1.25.0
opentelemetry-exporter-otlp-proto-grpc>=1.25.0
opentelemetry-exporter-otlp-proto-http>=1.30.0
opentelemetry-instrumentation>=0.46b0
opentelemetry-instrumentation-asgi>=0.46b0
opentelemetry-instrumentation-fastapi>=0.46b0
opentelemetry-proto>=1.25.0
opentelemetry-sdk>=1.30.0
opentelemetry-semantic-conventions>=0.46b0
opentelemetry-util-http>=0.46b0

# Authentication
oauthlib>=3.2.2

# Data Validation
pydantic>=2.4.2
pydantic_core>=2.8.0

# Utilities
protobuf>=4.25.3
python-dotenv>=1.0.0
pip>=24.0

# Database
sqlalchemy>=2.0.0

# Queue Worker
celery>=5.3.0
redis>=5.0.0

# PDF Processing
pypdf>=4.2.0